ai_model/requirements.txt
numpy>=2.0,<3.0
pillow>=10.4,<12.0
safetensors>=0.4,<1.0
torch>=2.3,<3.0
transformers>=4.44,<5.0
ultralytics>=8.3,<9.0
